Break CCD code into modules and replace aa-variants with patch
